EinsteinExact.m: Change scheduling to use ADMBase parameters
Previously, this thorn used its own parameters to determine whether to schedule in initial data or on every timestep.  Initial data was scheduled in ADMBase_InitialData.  This meant that the lapse was overwritten in ADMBase_InitialGauge by ADMBase as ADMBase was setting the lapse according to ADMBase::initial_lapse.

This patch modified the scheduling to use the ADMBase parameters initial_data and evolution_method.  This change is backward-incompatible - parameter files need to be changed.